[API] reduce memory footprint, double -> float for RTs and intensities

+++ src/openms/include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/MultiplexSatelliteProfile.h
@@ -56,7 +56,7 @@ namespace OpenMS
     /**
      * @brief returns the RT of the satellite data point
      */
-    double getRT() const;
+    float getRT() const;
     
     /**
      * @brief returns the m/z of the satellite data point
@@ -66,16 +66,16 @@ namespace OpenMS
     /**
      * @brief returns the intensity of the satellite data point
      */
-    double getIntensity() const;
+    float getIntensity() const;
     
     private:
      
     /**
      * @brief position and intensity of the data point within the spline-interpolated experiment
      */
-    double rt_;
+    float rt_;
     double mz_;
-    double intensity_;
+    float intensity_;
     
   };
